+++ model/notification.py
@@ -3,7 +3,7 @@ import datetime
 from typing import List
 from model.payment import payment
 from model.payment.payment import Payment
-from model.activity import Activity
+from model.timesheet.activity import Activity
 from util import amount as util_amount
 from util import date_time
